litter-api/requirements.txt
flask==3.1.0
flask-smorest==0.45.0
marshmallow==3.26.1
gunicorn==23.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
Pillow==11.1.0
numpy==2.2.3

# Add these in your real project once your runtime is pinned
# torch==...
# torchvision==...
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git@...